[libcxx][test] Silence signed/unsigned comparison warnings

+++ libcxx/test/std/containers/associative/set/iterator.pass.cpp
@@ -70,10 +70,10 @@ int main(int, char**)
         i = m.begin();
         std::set<int>::const_iterator k = i;
         assert(i == k);
-        for (int j = 1; static_cast<std::size_t>(j) <= m.size(); ++j, ++i)
+        for (int j = 1; j <= static_cast<int>(m.size()); ++j, ++i)
             assert(*i == j);
         assert(i == m.end());
-        for (int j = m.size(); j >= 1; --j) {
+        for (int j = static_cast<int>(m.size()); j >= 1; --j) {
             --i;
             assert(*i == j);
         }
@@ -115,10 +115,10 @@ int main(int, char**)
         assert(static_cast<std::size_t>(std::distance(m.crbegin(), m.crend())) == m.size());
         std::set<int>::const_iterator i;
         i = m.begin();
-        for (int j = 1; static_cast<std::size_t>(j) <= m.size(); ++j, ++i)
+        for (int j = 1; j <= static_cast<int>(m.size()); ++j, ++i)
             assert(*i == j);
         assert(i == m.end());
-        for (int j = m.size(); j >= 1; --j) {
+        for (int j = static_cast<int>(m.size()); j >= 1; --j) {
             --i;
             assert(*i == j);
         }
@@ -161,10 +161,10 @@ int main(int, char**)
         i = m.begin();
         std::set<int, std::less<int>, min_allocator<int>>::const_iterator k = i;
         assert(i == k);
-        for (int j = 1; static_cast<std::size_t>(j) <= m.size(); ++j, ++i)
+        for (int j = 1; j <= static_cast<int>(m.size()); ++j, ++i)
             assert(*i == j);
         assert(i == m.end());
-        for (int j = m.size(); j >= 1; --j) {
+        for (int j = static_cast<int>(m.size()); j >= 1; --j) {
             --i;
             assert(*i == j);
         }
@@ -206,10 +206,10 @@ int main(int, char**)
         assert(static_cast<std::size_t>(std::distance(m.crbegin(), m.crend())) == m.size());
         std::set<int, std::less<int>, min_allocator<int>>::const_iterator i;
         i = m.begin();
-        for (int j = 1; static_cast<std::size_t>(j) <= m.size(); ++j, ++i)
+        for (int j = 1; j <= static_cast<int>(m.size()); ++j, ++i)
             assert(*i == j);
         assert(i == m.end());
-        for (int j = m.size(); j >= 1; --j) {
+        for (int j = static_cast<int>(m.size()); j >= 1; --j) {
             --i;
             assert(*i == j);
         }
